DEV-1622 Swap values of startDate and endDate
The mail that is sent out also contains the amount of items that are going
to be published next month. The items are queried via the MediaHaven REST
API, with an archived date between 6 months and 5 months ago. However the
start date is 5 months ago and the end date is 6 months ago, meaning that
the start date is after the end date. This results in a 500 from MediaHaven
stopping the process.

Swap the values of startDate and endDate so a result is returned and the mail
is sent out.